reverts

+++ src/main/java/com/prat/student/model/SubjectRequest.java
@@ -10,15 +10,15 @@ public class SubjectRequest {
     private String subjectName;
 
     @NotNull(message = "Max Marks cannot be blank")
-    private Float maxMark;
+    private Integer maxMark;
 
     @NotNull(message = "Pass Marks cannot be blank")
-    private Float passMark;
+    private Integer passMark;
 
     @Min(1)
     private Integer maxAttempt;
 
-    public SubjectRequest( String subjectName, Float maxMark, Float passMark, Integer maxAttempt){
+    public SubjectRequest( String subjectName, Integer maxMark, Integer passMark, Integer maxAttempt){
         this.subjectName = subjectName;
         this.maxMark = maxMark;
         this.passMark = passMark;
